backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.116.1",
    "python-logstash>=0.4.8",
    "uvicorn>=0.35.0",
]
